+++ examples/with-angular-thirdpartyemailpassword/src/app/app.component.ts
@@ -1,9 +1,14 @@
 import { Component } from "@angular/core";
 
-import * as SuperTokens from "supertokens-website";
+import * as SuperTokens from "supertokens-web-js";
+import * as Session from "supertokens-web-js/recipe/session";
 
 SuperTokens.init({
-    apiDomain: "http://localhost:3001",
+    appInfo: {
+        appName: "SuperTokens Demo",
+        apiDomain: "http://localhost:3001",
+    },
+    recipeList: [Session.init()],
 });
 
 @Component({