+++ examples/with-multiple-email-sign-in/src/Home/index.tsx
@@ -1,0 +1,21 @@
+import React from "react";
+import Logout from "./Logout";
+import SuccessView from "./SuccessView";
+import { useNavigate } from "react-router-dom";
+import { signOut } from "supertokens-auth-react/recipe/emailpassword";
+
+export default function Home() {
+    const navigate = useNavigate();
+
+    async function logoutClicked() {
+        await signOut();
+        navigate("/auth");
+    }
+
+    return (
+        <div className="fill">
+            <Logout logoutClicked={logoutClicked} />
+            <SuccessView />
+        </div>
+    );
+}